Fix compilation errors and warnings (500+ issues resolved)
Fixed Issues:
1. Missing PacketTypes.h - Changed to NetworkProtocol.h
2. Color struct conflicts - Removed duplicate, use RenderEngine.h definition
3. Transform class/struct mismatch - Fixed forward declarations
4. Case label initialization - Added braces to WM_SIZE case
5. Double to float conversions - Using sqrtf() and explicit float literals
6. uint64_t to uint32_t conversions - Added static_cast for all sizeof()
7. Light constructor - Initialize all members including Color, Vec3, Mat4
8. Animation.h - Include RenderEngine.h for Color definition

Changes by file:
- LoginScene.cpp: Fix include, add uint32_t casts
- UIElement.h: Remove duplicate Color, include RenderEngine.h
- Animation.h: Include RenderEngine.h and GameObject.h, fix Transform
- main.cpp: Add braces to WM_SIZE case for halfAspect variable
- RenderEngine.h: Use sqrtf(), fix float literals, init Light members
- GameClient.cpp: Add GLdouble casts to glFrustum
- LobbyUI.cpp, LoginUI.cpp: Add uint32_t casts to sendPacket

All sendPacket calls now use static_cast<uint32_t>(sizeof(...)) to avoid
size_t to uint32_t conversion warnings.

+++ src/client/ui/GameClient.cpp
@@ -8,7 +8,9 @@
 static void myPerspective(double fovy, double aspect, double zNear, double zFar) {
     double fH = std::tan(fovy / 360.0 * 3.14159265) * zNear;
     double fW = fH * aspect;
-    glFrustum(-fW, fW, -fH, fH, zNear, zFar);
+    glFrustum(static_cast<GLdouble>(-fW), static_cast<GLdouble>(fW),
+              static_cast<GLdouble>(-fH), static_cast<GLdouble>(fH),
+              static_cast<GLdouble>(zNear), static_cast<GLdouble>(zFar));
 }
 
 GameClient::GameClient(NetworkClient* netClient, uint64_t accId)
@@ -1000,7 +1002,7 @@ void GameClient::sendPositionUpdate() {
     move.pitch = playerPitch;
     move.movementFlags = 0;
 
-    networkClient->sendPacket(PacketType::PLAYER_MOVE, &move, sizeof(move));
+    networkClient->sendPacket(PacketType::PLAYER_MOVE, &move, static_cast<uint32_t>(sizeof(move)));
 }
 
 void GameClient::handleSpawnInfo(const std::vector<uint8_t>& payload) {
